UNOMI-70 Upgrade to ElasticSearch 5.x - Working on getting integration testing working (not yet working)
