+++ itests/src/test/java/org/apache/unomi/itests/ConditionESQueryBuilderIT.java
@@ -14,40 +14,46 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
+
 package org.apache.unomi.itests;
 
-import org.apache.unomi.api.Profile;
-import org.apache.unomi.api.services.ProfileService;
-import org.junit.Test;
+import org.apache.unomi.api.Item;
+import org.apache.unomi.api.conditions.Condition;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.ops4j.pax.exam.junit.PaxExam;
 import org.ops4j.pax.exam.spi.reactors.ExamReactorStrategy;
 import org.ops4j.pax.exam.spi.reactors.PerSuite;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import javax.inject.Inject;
+import java.util.List;
 
 /**
- * An integration test for the profile service
+ * Integration tests for various condition query builder types (elasticsearch).
+ *
+ * @author Sergiy Shyrkov
  */
 @RunWith(PaxExam.class)
 @ExamReactorStrategy(PerSuite.class)
-public class ProfileServiceTest extends BaseTest {
+public class ConditionESQueryBuilderIT extends ConditionEvaluatorIT {
 
-    private final static Logger LOGGER = LoggerFactory.getLogger(SegmentTest.class);
-    private final static String TEST_PROFILE_ID = "test-profile-id";
-    @Inject
-    protected ProfileService profileService;
+    @Override
+    protected boolean eval(Condition c) {
+        @SuppressWarnings("unchecked")
+        List<Item> list = persistenceService.query(c,null,(Class<Item>) item.getClass());
+        return list.contains(item);
+    }
 
-    @Test
-    public void testProfileDelete() {
-        Profile profile = new Profile();
-        profile.setItemId(TEST_PROFILE_ID);
-        profileService.save(profile);
-        LOGGER.info("Profile saved, now testing profile delete...");
-        profileService.delete(TEST_PROFILE_ID, false);
-        LOGGER.info("Profile deleted successfully.");
+    @Before
+    public void setUp() {
+        super.setUp();
+        persistenceService.save(item);
+        persistenceService.refresh();
+    }
+
+    @After
+    public void tearDown() {
+        persistenceService.remove(item.getItemId(), item.getClass());
     }
 
 }
